% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_dcdh_decomp.R
\name{compute_dcdh_weights}
\alias{compute_dcdh_weights}
\title{Produce a table with de Chaisemartin and D'Haultfoeuille weights}
\usage{
compute_dcdh_weights(data)
}
\arguments{
\item{data}{A table with:
\itemize{
\item \code{group}: id of the group (Primary key 1).
\item \code{t}: time (Primary key 2).
\item \code{size}: size of the group.
\item \code{treated}: treatment value.
}}
}
\value{
A table with the same rows as \code{data} and with:
\itemize{
\item \code{group}: (Primary key 1).
\item \code{t}: (Primary key 2).
\item \code{treated_group}: mean treatment of group.
\item \code{treated_t}: mean treatment at time \code{t}.
\item \code{treated_mean}: global mean treatment.
\item \code{eps}: residual of double demeaning.
\item \code{w}: weight in the de Chaisemartin and D'Haultfoeuille decomposition.
}
}
\description{
Produce a table with de Chaisemartin and D'Haultfoeuille weights
}
\examples{
data_ex <- data.frame(
  group = rep(c("control", "treated_1"), each = 10),
  t = c(1:10, 1:10),
  size = rep(c(1, 1), each = 10),
  treated = c(rep(0, 10), c(rep(0, 5), rep(1, 5)))
)
compute_dcdh_weights(data_ex)
}
