% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_goodman_bacon_decomp.R
\name{did_estimate}
\alias{did_estimate}
\title{Basic Difference in Difference estimation around an event}
\usage{
did_estimate(data, group_control, group_treated, event)
}
\arguments{
\item{data}{Table containing:
\itemize{
\item \code{group}: name of the group.
\item \code{t}: date.
\item \code{y}: outcome.
}}

\item{group_control}{Name of the \code{control} group.}

\item{group_treated}{Name of the \code{treated} group.}

\item{event}{Date of the treatment event.}
}
\value{
The Difference in Difference estimation of the treatment effect.
}
\description{
Basic Difference in Difference estimation around an event
}
\examples{
data <- data.frame(
  t = c(1:10, 1:10),
  group = rep(c("group_1", "group_2"), each = 10),
  y = c(1:10, c(1:4, 6:11))
)
did_estimate(data, "group_1", "group_2", 5)
}
