% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_goodman_bacon_decomp.R
\name{goodman_bacon_coef}
\alias{goodman_bacon_coef}
\title{Goodman-Bacon Difference in Difference weights}
\usage{
goodman_bacon_coef(nb_times, group_1, group_2)
}
\arguments{
\item{nb_times}{Number of time observations.}

\item{group_1}{List corresponding to the first group and containing:
\itemize{
\item \code{size}: size of the group.
\item \code{event}: time of the treatment event (eventually infinite for a control).
}}

\item{group_2}{List corresponding to the second group and containing:
\itemize{
\item \code{size}: size of the group.
\item \code{event}: time of the treatment event (eventually infinite for a control).
}}
}
\value{
A list containing the non-normalized DiD weights in the Goodman-Bacon
decomposition for the DiD estimators:
\itemize{
\item \code{s_12_1}: using \code{group_1} as the treated and \code{group_2} as the control.
\item \code{s_12_2}: using \code{group_2} as the treated and \code{group_1} as the control.
}
}
\description{
Goodman-Bacon Difference in Difference weights
}
\examples{
goodman_bacon_coef(11, list(size = 5, event = 6), list(size = 1, event = 3))
}
