% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_goodman_bacon_decomp.R
\name{goodman_bacon_decomp_params}
\alias{goodman_bacon_decomp_params}
\title{Goodman-Bacon Difference in Difference weights}
\usage{
goodman_bacon_decomp_params(nb_times, group_1, group_2)
}
\arguments{
\item{nb_times}{Number of time observations.}

\item{group_1}{List corresponding to the first group and containing:
\itemize{
\item \code{size}: size of the group.
\item \code{event}: time of the treatment event (eventually infinite for a control).
}}

\item{group_2}{List corresponding to the second group and containing:
\itemize{
\item \code{size}: size of the group.
\item \code{event}: time of the treatment event (eventually infinite for a control).
}}
}
\value{
A list containing the various parameters in the Goodman-Bacon
decomposition for the DiD estimators:
\itemize{
\item \code{d_1}: share of treatment time of \code{group_1}.
\item \code{d_2}: share of treatment time of \code{group_2}.
\item \code{n_12_1}: pure size effect of the DiD using \code{group_1} as the treated and \code{group_2} as the control.
\item \code{n_12_2}: pure size effect of the DiD using \code{group_2} as the treated and \code{group_1} as the control.
\item \code{v_n_12}: variance of the share of size between \code{group_1} and \code{group_2}.
\item \code{v_d_12_1}: variance of the share of treatment time in the DiD using \code{group_1} as the treated and \code{group_2} as the control.
\item \code{v_d_12_2}: variance of the share of treatment time in the DiD using \code{group_2} as the treated and \code{group_1} as the control.
\item \code{s_12_1}: unnormalized DiD weights using \code{group_1} as the treated and \code{group_2} as the control.
\item \code{s_12_2}: unnormalized DiD weights using \code{group_2} as the treated and \code{group_1} as the control.
}
}
\description{
Goodman-Bacon Difference in Difference weights
}
\examples{
goodman_bacon_decomp_params(11, list(size = 5, event = 6), list(size = 1, event = 3))
}
