% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_utils.R
\name{insert_name}
\alias{insert_name}
\title{Insert the name of each sublist of a list inside it}
\usage{
insert_name(x)
}
\arguments{
\item{x}{A list of list.}
}
\value{
A list of list, where each sublist contains a \code{name} attribute
corresponding to its name.
}
\description{
Insert the name of each sublist of a list inside it
}
\examples{
insert_name(list(a = list(aa = 1, ab = 2), b = list()))
}
