% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_math.R
\name{min_max}
\alias{min_max}
\title{Clamping values between two bounds}
\usage{
min_max(x, mini = 0, maxi = 1)
}
\arguments{
\item{x}{A vector to clamp.}

\item{mini}{Lower bound.}

\item{maxi}{Upper bound.}
}
\value{
A vector whose values are between the two bounds.
}
\description{
Clamping values between two bounds
}
\examples{
min_max(c(-5, 1, 2, 10), 0, 5)
}
