% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_prettify.R
\name{pp_col}
\alias{pp_col}
\title{Pretty relabeling of some values}
\usage{
pp_col(x)
}
\arguments{
\item{x}{A vector with names to prettify.}
}
\value{
A vector with prettified names.
}
\description{
Pretty relabeling of some values
}
\examples{
pp_col(c("control_1", "treated_2_12"))
}
