% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twfeBenchmark-package.R
\docType{package}
\name{twfeBenchmark}
\alias{twfeBenchmark-package}
\alias{twfeBenchmark}
\title{Shiny Application Visualizing Two-Way Fixed Effect Estimations}
\description{
This is an exploration tool allowing to understand the underlying assumptions
behind difference in difference estimators, two-way fixed estimators, as well
as the recent estimators and decomposition proposed by the literature in this
subject.
}
\keyword{internal}
