% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_prettify.R
\name{unpp_col}
\alias{unpp_col}
\title{Recover the logic values associated to a pretty name}
\usage{
unpp_col(x)
}
\arguments{
\item{x}{A vector with pretty names.}
}
\value{
A vector with unprettified names.
}
\description{
Recover the logic values associated to a pretty name
}
\examples{
unpp_col(c("Never Treated 1", "Treated 2 12"))
}
