% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_twfe_app}
\alias{run_twfe_app}
\title{Run the main shiny application}
\usage{
run_twfe_app(...)
}
\arguments{
\item{...}{}
}
\description{
Run the main shiny application
}
\examples{
\dontrun{run_twfe_app()}
}
